Refactor viz to it's own submodule
To avoid dependency bloating the signal-processing module waveforms, waveforms.viz
has been moved to the submodule waveforms-viz, which can be installed as an extra
`pip install waveforms[viz]`.
Support for submodule with relative path has only been implemented (well) in
poetry and hatchling.
For this reason, replace build backend with hatchling.
These changes only affect build and installation, the viz extra will still share
the waveforms namespace, e.g. `from waveforms.viz import eye_diagram`.

Changes:
- Build backend: setuptools -> hatchling
- filter -> filters to avoid conflict with python builtin
- waveforms/viz moved to waveforms-viz/waveforms/viz
- waveforms-viz pyproject.toml added.
- Added constellation plot.
- Added PySide6 and pyqtgraph as a viz dependency.

+++ waveforms-viz/waveforms/viz/eye.py
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
 
 def eye_diagram(  # noqa: PLR0913
     time: NDArray[np.float64],
-    signal: NDArray[np.float64],
+    signal: NDArray[np.complex64],
     sps: int = 8,
     modulo: int = 4,
     *,
